% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{coef.emmil_model}
\alias{coef.emmil_model}
\title{Find an EMMIL model's coefficients}
\usage{
\method{coef}{emmil_model}(object, ...)
}
\arguments{
\item{object}{TO DO}

\item{...}{TO DO}

\item{adjust}{TO DO}
}
\value{
TO DO
}
\description{
Find an EMMIL model's coefficients
}
\examples{
NULL

}
