% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmil_model.R
\name{coef.mmil_model}
\alias{coef.mmil_model}
\title{Find an MMIL model's coefficients}
\usage{
\method{coef}{mmil_model}(object, ...)
}
\arguments{
\item{object}{An MMIL model.}

\item{...}{Additional arguments to pass to the subclass-specific `coef` function.}
}
\value{
TODO - Timothy
}
\description{
Find an MMIL model's coefficients
}
\examples{
NULL

}
