% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_calculate_case_control_adjustment}
\alias{emmil_calculate_case_control_adjustment}
\title{Title}
\usage{
emmil_calculate_case_control_adjustment(z, rho, zeta)
}
\arguments{
\item{z}{A numeric vector of inherited patient labels for each cell. 1 if the 
cell was sampled from a patient with the disease and 0 otherwise.}

\item{rho}{A scalar value between 0 and 1 indicating the probability that 
that a cell is NOT associated with the outcome-of-interest, given that the 
cell was sampled from a patient who DOES have the outcome of interest. In other 
words, the expected number of non-disease associated cells in each sick patient.}

\item{zeta}{A scalar value between 0 and 1 indicating the marginal probability that an 
inherited patient label will be 1. In other words, something like the probabilty 
that a random person from your population-of-interest will have the condition-of-interest
(TODO: ASK ERIN ABOUT THIS).}
}
\value{
A scalar value indicating the case control adjustment that must be 
added to our logit to adjust the intercept of our model.
}
\description{
Title
}
\examples{
NULL

}
