% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_calculate_sample_label_adjustment}
\alias{emmil_calculate_sample_label_adjustment}
\title{Calculate the sample-label adjustment in an EMMIL model.}
\usage{
emmil_calculate_sample_label_adjustment(rho, zeta)
}
\arguments{
\item{rho}{A scalar value between 0 and 1 indicating the probability that
that a cell is NOT associated with the outcome-of-interest, given that the
cell was sampled from a patient who DOES have the outcome of interest. In other
words, the expected number of non-disease-associated cells in each sick patient.}

\item{zeta}{A scalar value between 0 and 1 indicating the marginal probability that an
inherited patient label will be 1. This is the probability
that a random cell from your population-of-interest will be sampled from a
person who has the condition-of-interest.}
}
\value{
A scalar value indicating the case control adjustment that must be
added to the predicted logit to effectively adjust the intercept of the model.
}
\description{
This function calculates the sample-label adjustment used during EMMIL
modeling fitting and prediction. In brief, this adjustment adjusts the probability
scores for cells sampled from sick patients (i.e. those for whom z = 1), for whom
we expect a higher degree of disease-association.
}
\examples{
NULL

}
