% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_fit_glm}
\alias{emmil_fit_glm}
\title{Fit an EMMIL model using a generalized linear model.}
\usage{
emmil_fit_glm(
  X,
  z,
  rho,
  zeta,
  case_control_adjustment = 0,
  num_iterations = 20L
)
}
\arguments{
\item{X}{TODO}

\item{z}{TODO}

\item{rho}{TODO}

\item{zeta}{TODO}

\item{case_control_adjustment}{TODO}

\item{num_iterations}{TODO}

\item{lambda}{TODO}
}
\value{
A list.
}
\description{
Fit an EMMIL model using a generalized linear model.
}
\examples{
NULL

}
