% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_fit_mlp}
\alias{emmil_fit_mlp}
\title{Fit an EMMIL model using a multilayer perceptron.}
\usage{
emmil_fit_mlp(
  X,
  z,
  rho,
  zeta,
  case_control_adjustment = 0,
  num_iterations = 20L,
  num_neurons,
  decay,
  ...
)
}
\arguments{
\item{X}{TO DO}

\item{z}{TO DO}

\item{rho}{TO DO}

\item{zeta}{TO DO}

\item{case_control_adjustment}{TO DO}

\item{num_iterations}{TO DO}

\item{num_neurons}{TO DO}

\item{decay}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
Fit an EMMIL model using a multilayer perceptron.
}
\examples{
NULL
}
