% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_fit_tidymodels}
\alias{emmil_fit_tidymodels}
\title{Title}
\usage{
emmil_fit_tidymodels(
  X,
  z,
  rho,
  zeta,
  case_control_adjustment = 0,
  num_iterations = 20L,
  model_spec
)
}
\arguments{
\item{X}{TO DO}

\item{z}{TO DO}

\item{rho}{TO DO}

\item{zeta}{TO DO}

\item{case_control_adjustment}{TO DO}

\item{num_iterations}{TO DO}

\item{model_spec}{Any tidymodels model specification.}
}
\value{
TO DO
}
\description{
Title
}
\examples{
NULL
}
