% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{emmil_get_hyperparameters}
\alias{emmil_get_hyperparameters}
\title{Get the hyperparameters of an EMMIL object}
\usage{
emmil_get_hyperparameters(emmil_model_object)
}
\arguments{
\item{emmil_model_object}{An EMMIL model.}
}
\value{
A list of model type-specific hyperparameters
}
\description{
Get the hyperparameters of an EMMIL object
}
\examples{
NULL
}
