% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{emmil_get_log_likelihoods}
\alias{emmil_get_log_likelihoods}
\title{Get the training log-likelihood values of an EMMIL object}
\usage{
emmil_get_log_likelihoods(emmil_model_object)
}
\arguments{
\item{emmil_model_object}{An EMMIL model.}
}
\value{
A vector of log-likelihood values (one for each iteration)
}
\description{
Get the training log-likelihood values of an EMMIL object
}
\examples{
NULL
}
