% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{emmil_loglik}
\alias{emmil_loglik}
\title{Compute the log-likelihood of the EM-MIL algorithm}
\usage{
emmil_loglik(z, predictions, rho, num_cells_disease, num_cells)
}
\arguments{
\item{z}{An integer vector containing the inherited (parent) labels for each cell}

\item{predictions}{A numeric vector containing the (logit) model predictions from the maximization step}

\item{rho}{The probability that a cell from a diseased sample is not 
disease-associated (i.e. the probability that y = 0 when z = 1).}

\item{num_cells_disease}{A scalar indicating the number of cells in the (training) dataset that come 
from samples with the disease.}

\item{num_cells}{A scalar indicating the total number of cells in the (training) dataset.}
}
\value{
A numeric scalar indicating the log-likelihood for the current iteration.
}
\description{
Compute the log-likelihood of the EM-MIL algorithm
}
