% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{emmil_model}
\alias{emmil_model}
\title{Construct an EMMIL model.}
\usage{
emmil_model(
  model,
  rho = 0.5,
  zeta = 0.5,
  num_cells,
  num_cells_disease,
  lls = 0,
  hyperparameters = list()
)
}
\arguments{
\item{model}{TO DO}

\item{rho}{TO DO}

\item{zeta}{TO DO}

\item{num_cells}{TO DO}

\item{num_cells_disease}{TO DO}

\item{lls}{TO DO}

\item{hyperparameters}{TO DO}
}
\value{
TO DO
}
\description{
Construct an EMMIL model.
}
\seealso{
Other emmil_model utilities: 
\code{\link{new_emmil_model}()}
}
\concept{emmil_model utilities}
