% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_plot_lls}
\alias{emmil_plot_lls}
\title{Plot the log-likelihood values over each iteration of EMMIL.}
\usage{
emmil_plot_lls(emmil_model_object, lls, theme = ggplot2::theme_bw(), ...)
}
\arguments{
\item{emmil_model_object}{An EMMIL model fit using the high-level API.}

\item{lls}{A numeric vector of log-likelihood values. Will be ignored if
emmil_model_object is provided.}

\item{theme}{A ggplot2 theme. Defaults to ggplot2::theme_bw().}

\item{...}{Optional additional arguments to pass to ggplot2::geom_point().}
}
\value{
A ggplot2 object.
}
\description{
Plot the log-likelihood values over each iteration of EMMIL.
}
\examples{
NULL

}
