% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_simulate_data}
\alias{emmil_simulate_data}
\title{Simulate data to demonstrate how to use the EM-MIL model.}
\usage{
emmil_simulate_data(
  num_cells = 100,
  num_features = 10,
  rho = 0.5,
  zeta = 0.5,
  shift = 4
)
}
\arguments{
\item{num_cells}{An integer number of cells to simulate}

\item{num_features}{An integer number of features to simulate}

\item{rho}{A scalar value between 0 and 1 indicating the probability that 
that a cell is NOT associated with the outcome-of-interest, given that the 
cell was sampled from a patient who DOES have the outcome of interest. In other 
words, the expected number of non-disease associated cells in each sick patient.}

\item{zeta}{A scalar value between 0 and 1 indicating the marginal probability that an 
inherited patient label will be 1. In other words, something like the probabilty 
that a random person from your population-of-interest will have the condition-of-interest
(TODO: ASK ERIN ABOUT THIS).}

\item{shift}{}
}
\value{
A list with 3 items: X (the expression matrix), z (a numeric 
vector of inherited patient labels for each cell) and true_y (an integer vector 
indicating the "true" answer for whether or not a cell is associated with
the outcome-of-interest (1) or not (0)).
}
\description{
Simulate data to demonstrate how to use the EM-MIL model.
}
\examples{
NULL
}
