% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initialize_y}
\alias{initialize_y}
\title{A function that initializes y_0 (the y vector for the 0th iteration) for the EM-MIL algorithm.}
\usage{
initialize_y(z, rho)
}
\arguments{
\item{z}{The vector of inherited parent labels for each cell.}

\item{rho}{The probability that a cell from a diseased sample is not 
disease-associated (i.e. the probability that y = 0 when z = 1).}
}
\value{
A vector of the same length as z representing the initial disease-
association probabilities for each cell in the dataset.
}
\description{
A function that initializes y_0 (the y vector for the 0th iteration) for the EM-MIL algorithm.
}
