% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{logit}
\alias{logit}
\title{The logit function.}
\usage{
logit(p)
}
\arguments{
\item{p}{A vector of numeric scores (generally, probabilities between 0 and 1).}
}
\value{
A vector of the same length as `p`, after logit transformation.
}
\description{
The logit function.
}
\examples{
NULL

}
