% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmil_model.R
\name{mmil_get_hyperparameters}
\alias{mmil_get_hyperparameters}
\title{Get the hyperparameters of an MMIL object}
\usage{
mmil_get_hyperparameters(mmil_model_object)
}
\arguments{
\item{mmil_model_object}{An MMIL model.}
}
\value{
A list of model type-specific hyperparameters
}
\description{
Get the hyperparameters of an MMIL object
}
\examples{
NULL
}
