% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmil_model.R
\name{mmil_get_log_likelihoods}
\alias{mmil_get_log_likelihoods}
\title{Get the training log-likelihood values of an MMIL object}
\usage{
mmil_get_log_likelihoods(mmil_model_object)
}
\arguments{
\item{mmil_model_object}{An MMIL model.}
}
\value{
A vector of log-likelihood values (one for each iteration)
}
\description{
Get the training log-likelihood values of an MMIL object
}
\examples{
NULL
}
