% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmil_model.R
\name{mmil_model}
\alias{mmil_model}
\title{Construct an MMIL model.}
\usage{
mmil_model(
  model,
  rho = 0.5,
  zeta = 0.5,
  num_cells,
  num_cells_disease,
  lls = 0,
  hyperparameters = list()
)
}
\arguments{
\item{model}{TO DO}

\item{rho}{TO DO}

\item{zeta}{TO DO}

\item{num_cells}{TO DO}

\item{num_cells_disease}{TO DO}

\item{lls}{TO DO}

\item{hyperparameters}{TO DO}
}
\value{
TO DO
}
\description{
Construct an MMIL model.
}
\seealso{
Other mmil_model utilities: 
\code{\link{new_mmil_model}()}
}
\concept{mmil_model utilities}
