% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{mmil_plot_lls}
\alias{mmil_plot_lls}
\title{Plot the log-likelihood values over each iteration of MMIL.}
\usage{
mmil_plot_lls(mmil_model_object, lls, theme = ggplot2::theme_bw(), ...)
}
\arguments{
\item{mmil_model_object}{An MMIL model fit using the high-level API.}

\item{lls}{A numeric vector of log-likelihood values. Will be ignored if
mmil_model_object is provided.}

\item{theme}{A ggplot2 theme. Defaults to ggplot2::theme_bw().}

\item{...}{Optional additional arguments to pass to ggplot2::geom_point().}
}
\value{
A ggplot2 object.
}
\description{
Plot the log-likelihood values over each iteration of MMIL.
}
\examples{
NULL

}
