% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{new_emmil_model}
\alias{new_emmil_model}
\title{Construct an EMMIL model.}
\usage{
new_emmil_model(
  model,
  rho,
  zeta,
  num_cells,
  num_cells_disease,
  lls,
  hyperparameters = list()
)
}
\arguments{
\item{model}{TO DO}

\item{rho}{TO DO}

\item{zeta}{TO DO}

\item{num_cells}{TO DO}

\item{num_cells_disease}{TO DO}

\item{lls}{TO DO}

\item{hyperparameters}{TO DO}
}
\value{
An "emmil_model", an S3 object subclass of the original model class. 
Attributes include rho, zeta, num_cells, num_cells_disease, and hyperparameters.
}
\description{
Construct an EMMIL model.
}
\seealso{
Other emmil_model utilities: 
\code{\link{emmil_model}()}
}
\concept{emmil_model utilities}
