% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmil_model.R
\name{new_mmil_model}
\alias{new_mmil_model}
\title{Construct an MMIL model.}
\usage{
new_mmil_model(
  model,
  rho,
  zeta,
  num_cells,
  num_cells_disease,
  lls,
  hyperparameters = list()
)
}
\arguments{
\item{model}{TO DO}

\item{rho}{A scalar value between 0 and 1 indicating the probability that
that a cell is NOT associated with the outcome-of-interest, given that the
cell was sampled from a patient who DOES have the outcome of interest. In other
words, the expected number of non-disease-associated cells in each sick patient.}

\item{zeta}{The probability that the inherited patient label of any given cell will be 1.}

\item{num_cells}{TO DO}

\item{num_cells_disease}{TO DO}

\item{lls}{TO DO}

\item{hyperparameters}{TO DO}
}
\value{
An "mmil_model", an S3 object subclass of the original model class. 
Attributes include rho, zeta, num_cells, num_cells_disease, and hyperparameters.
}
\description{
Construct an MMIL model.
}
\seealso{
Other mmil_model utilities: 
\code{\link{mmil_model}()}
}
\concept{mmil_model utilities}
