% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{p_from_logit}
\alias{p_from_logit}
\title{Obtain a binary class probability from a logit}
\usage{
p_from_logit(logit)
}
\arguments{
\item{logit}{A numeric vector of logit (log-odds) values}
}
\value{
A numeric vector of the same length as logit representing the probability
that each observation is of class 1.
}
\description{
Obtain a binary class probability from a logit
}
