% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmil_model.R
\name{predict.mmil_model}
\alias{predict.mmil_model}
\title{Make predictions for new data using an MMIL model.}
\usage{
\method{predict}{mmil_model}(object, ...)
}
\arguments{
\item{object}{An mmil_model object.}

\item{...}{Additional parameters to pass to the object subclass.}
}
\value{
Predictions based on the subclass predict method
}
\description{
Make predictions for new data using an MMIL model.
}
\examples{
NULL

}
