% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{coef.emmil_model}
\alias{coef.emmil_model}
\title{Find an EMMIL model's coefficients}
\usage{
\method{coef}{emmil_model}(object, ...)
}
\arguments{
\item{object}{An EMMIL model.}

\item{...}{Additional arguments to pass to the subclass-specific `coef` function.}
}
\value{
TODO - Timothy
}
\description{
Find an EMMIL model's coefficients
}
\examples{
NULL

}
