% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_fit_mlp}
\alias{emmil_fit_mlp}
\title{Fit an EMMIL model using a multilayer perceptron.}
\usage{
emmil_fit_mlp(
  X,
  z,
  rho,
  zeta,
  case_control_adjustment = 0,
  num_iterations = 20L,
  num_neurons,
  decay,
  early_stopping_patience = 3L,
  early_stopping_tolerance = 1e-04,
  ...
)
}
\arguments{
\item{X}{The numeric feature matrix of single-cell data. Each row is a cell, and each
column is a feature.}

\item{z}{A numeric vector of inherited patient labels for each cell. 1 if the
cell was sampled from a patient with the disease and 0 otherwise.}

\item{rho}{A scalar value between 0 and 1 indicating the probability that
that a cell is NOT associated with the outcome-of-interest, given that the
cell was sampled from a patient who DOES have the outcome of interest. In other
words, the expected number of non-disease-associated cells in each sick patient.}

\item{zeta}{A scalar value between 0 and 1 indicating the marginal probability that an
inherited patient label will be 1. In other words, the probability
that a random cell from your population-of-interest will be from a patient with
the condition-of-interest.}

\item{case_control_adjustment}{TO DO}

\item{num_iterations}{An integer indicating the maximum
number of iterations to use when fitting the model.}

\item{num_neurons}{TO DO}

\item{decay}{TO DO}

\item{early_stopping_patience}{An integer number of iterations for which the improvement
in the log-likelihood must be lower than `early_stopping_tolerance` in a row before
model training is stopped. Defaults to 3. Must be larger than 0.}

\item{early_stopping_tolerance}{A scalar value indicating the minimum value by which the
training log-likelihood must improve to continue training. If the improvement in the
log-likelihood is smaller than this value for `early_stopping_patience` iterations
in a row, model training will be stopped before `num_iterations` iterations are fit.
Defaults to 1e-4. This value is for the normalized log-likelihood (i.e. divided by number of cells in the training set).}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
Fit an EMMIL model using a multilayer perceptron.
}
\examples{
NULL
}
