% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{emmil_fit_tidymodels}
\alias{emmil_fit_tidymodels}
\title{Title}
\usage{
emmil_fit_tidymodels(
  X,
  z,
  rho,
  zeta,
  case_control_adjustment = 0,
  num_iterations = 20L,
  model_spec
)
}
\arguments{
\item{X}{TO DO}

\item{z}{A numeric vector of inherited patient labels for each cell. 1 if the
cell was sampled from a patient with the disease and 0 otherwise.}

\item{rho}{A scalar value between 0 and 1 indicating the probability that
that a cell is NOT associated with the outcome-of-interest, given that the
cell was sampled from a patient who DOES have the outcome of interest. In other
words, the expected number of non-disease-associated cells in each sick patient.}

\item{zeta}{A scalar value between 0 and 1 indicating the marginal probability that an
inherited patient label will be 1. In other words, the probability
that a random cell from your population-of-interest will be from a patient with
the condition-of-interest.}

\item{case_control_adjustment}{TO DO}

\item{num_iterations}{TO DO}

\item{model_spec}{Any tidymodels model specification.}
}
\value{
TO DO
}
\description{
Title
}
\examples{
NULL
}
