% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{predict.emmil_model}
\alias{predict.emmil_model}
\title{Make predictions for new data using an EMMIL model.}
\usage{
\method{predict}{emmil_model}(object, ...)
}
\arguments{
\item{object}{TO DO}

\item{...}{TO DO}
}
\value{
TO DO
}
\description{
Make predictions for new data using an EMMIL model.
}
\examples{
NULL

}
