% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmil_model.R
\name{predict.emmil_model}
\alias{predict.emmil_model}
\title{Make predictions for new data using an EMMIL model.}
\usage{
\method{predict}{emmil_model}(object, ...)
}
\arguments{
\item{object}{An emmil_model object.}

\item{...}{Additional parameters to pass to the object subclass.}
}
\value{
Predictions based on the subclass predict method
}
\description{
Make predictions for new data using an EMMIL model.
}
\examples{
NULL

}
